services/control-plane/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4
supabase==2.7.4
anthropic==0.34.2
openai==1.45.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.8
